requirements.txt
streamlit
google-generativeai
python-dotenv
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
youtube_transcript_api